# openapi.json
{"openapi":"3.1.0","info":{"title":"Flock Energy API","description":"A clean REST API wrapper around the legacy Urja Meter Ops portal.","version":"1.0.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Force a (re-)login against the legacy portal.","operationId":"login_api_v1_auth_login_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/v1/meters":{"get":{"tags":["meters"],"summary":"List Meters","operationId":"list_meters_api_v1_meters_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number, starting at 1","default":1,"title":"Page"},"description":"Page number, starting at 1"},{"name":"q","in":"query","required":false,"schema":{"type":"string","description":"Search by meter number or serial","default":"","title":"Q"},"description":"Search by meter number or serial"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeterListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}":{"get":{"tags":["meters"],"summary":"Get Meter","operationId":"get_meter_api_v1_meters__meter_id__get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeterDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}/consumption":{"get":{"tags":["meters"],"summary":"Get Consumption","operationId":"get_consumption_api_v1_meters__meter_id__consumption_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsumptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["meta"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ConsumptionReading":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"kwh":{"type":"number","title":"Kwh"},"kvah":{"type":"number","title":"Kvah"},"voltR":{"type":"number","title":"Voltr"}},"type":"object","required":["timestamp","kwh","kvah","voltR"],"title":"ConsumptionReading","description":"One normalized consumption data point.\n\nLegacy format: {\"timestamp\": \"23/06/2026 23:30\", \"kwh\": \"48438.74\", ...}\nNormalized: real datetime object, real floats."},"ConsumptionResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"readings":{"items":{"$ref":"#/components/schemas/ConsumptionReading"},"type":"array","title":"Readings"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["meter_id","readings","count"],"title":"ConsumptionResponse"},"GeoLocation":{"properties":{"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"}},"type":"object","required":["latitude","longitude"],"title":"GeoLocation"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginResponse":{"properties":{"authenticated":{"type":"boolean","title":"Authenticated"},"logged_in_at":{"type":"string","format":"date-time","title":"Logged In At"}},"type":"object","required":["authenticated","logged_in_at"],"title":"LoginResponse"},"MeterDetail":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"serial_no":{"type":"string","title":"Serial No"},"make":{"type":"string","title":"Make"},"phase_type":{"type":"string","title":"Phase Type"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"dt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dt"},"location":{"anyOf":[{"$ref":"#/components/schemas/GeoLocation"},{"type":"null"}]}},"type":"object","required":["meter_id","serial_no","make","phase_type"],"title":"MeterDetail"},"MeterListResponse":{"properties":{"data":{"items":{"$ref":"#/components/schemas/MeterSummary"},"type":"array","title":"Data"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"}},"type":"object","required":["data","total","page","page_size"],"title":"MeterListResponse","description":"Our API's response shape for GET /api/v1/meters."},"MeterSummary":{"properties":{"meterId":{"type":"string","title":"Meterid"},"serialNo":{"type":"string","title":"Serialno"},"make":{"type":"string","title":"Make"},"phaseType":{"type":"string","title":"Phasetype"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"dt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dt"}},"type":"object","required":["meterId","serialNo","make","phaseType"],"title":"MeterSummary","description":"A single meter as returned by the legacy /portal/meters/search endpoint."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}